// frontend/src/types/Impact/ads.ts
import { AdType } from "./enums/AdType";
import { BogoBuyScope } from "./enums/BogoBuyScope";
import { BogoGetDiscountType } from "./enums/BogoGetDiscountType";
import { DealScope } from "./enums/DealScope";
import { DealState } from "./enums/DealState";
import { DealType } from "./enums/DealType";
import { Languages } from "./enums/Languages";
import { Season } from "./enums/Seasons";

export type TAd = {
    Id: string;
    Name: string;
    Description: string;
    CampaignId?: string;
    CampaignName: string;
    Type: AdType;
    TrackingLink: string;
    LandingPageUrl: string;
    AdvertiserId: string;
    AdvertiserName: string;
    Code: string;
    IFrameCode: string;
    Width: string;
    Height: string;
    CreativeUrl: string;
    Labels: string;
    AllowDeepLinking: string;
    MobileReady?: string;
    Language: Languages;
    StartDate: string;
    EndDate: string;
    Season: Season;
    TopSeller: string;
    SubjectLines: string;
    FromAddresses: string;
    UnsubscribeLink: string;
    DealId?: number;
    DealName: string;
    DealDescription: string;
    DealState: DealState;
    DealType: DealType;
    DealScope: DealScope;
    DealPublic: string;
    DealProducts: string[]; // Assuming it's an array of strings
    DealCategories: string;
    DiscountType: string;
    DiscountAmount: string;
    DiscountCurrency: string;
    DiscountPercent: string;
    DiscountMaximumPercent: string;
    DiscountPercentRangeStart: string;
    DiscountPercentRangeEnd: string;
    Gift: string;
    RebateAmount: string;
    RebateCurrency: string;
    DealDefaultPromoCode: string;
    MinimumPurchaseAmount: string;
    MinimumPurchaseAmountCurrency: string;
    MaximumSavingsAmount: string;
    MaximumSavingsCurrency: string;
    BogoBuyQuantity: string;
    BogoBuyScope: BogoBuyScope;
    BogoBuyName: string;
    BogoBuyImageUrl: string;
    BogoGetQuantity: string;
    BogoGetScope: BogoBuyScope; // same at that of BogoBuyScope
    BogoGetDiscountType: BogoGetDiscountType;
    BogoGetName: string;
    BogoGetImageUrl: string;
    BogoGetDiscountAmount: string;
    BogoGetDiscountCurrency: string;
    BogoGetDiscountPercent: string;
    PurchaseLimitQuantity: string;
    DealStartDate: string;
    DealEndDate: string;
    DealDateCreated: string;
    DealDateLastUpdated: string;
    Uri: string;
};